замена типа суммы int -> long

+++ Program.cs
@@ -14,25 +14,26 @@ static void RunTest(int[] array)
 {
     Stopwatch stopwatch = new();
     stopwatch.Start();
-    int sum1 = SumArray.RunSum(array);
-   
+    long sum1 = SumArray.RunSum(array);
+
     stopwatch.Stop();
     long time1 = stopwatch.ElapsedTicks;
     long time11 = stopwatch.ElapsedMilliseconds;
     Console.WriteLine($"Обычный: {time1} ticks, {time11} mc");
-    
+
     for (int i = 2; i < 9; i += 2)
-    {   if (i == 6) continue;
+    {
+        if (i == 6) continue;
         stopwatch.Restart();
-        int sum2 = SumArrayThread.RunSum(array,i );
+        long sum2 = SumArrayThread.RunSum(array, i);
         stopwatch.Stop();
         long time2 = stopwatch.ElapsedTicks;
         long time21 = stopwatch.ElapsedMilliseconds;
         Console.WriteLine($"Параллельный Thread потоков {i}:{time2} ticks, {time21} mc");
     }
-    
+
     stopwatch.Restart();
-    int sum3 = SumArrayParallelLINQ.RunSum(array);
+    long sum3 = SumArrayParallelLINQ.RunSum(array);
     stopwatch.Stop();
     long time3 = stopwatch.ElapsedTicks;
     long time31 = stopwatch.ElapsedMilliseconds;
