+++ TestSum/SumArray.cs
@@ -8,9 +8,9 @@ namespace DZ_ParallelLINQ.TestSum
 {
     public class SumArray()
     {
-        public static int RunSum(int[] array)
+        public static long RunSum(int[] array)
         {
-            int sum = 0;
+            long sum = 0;
             for (int i = 0; i < array.Length; i++)
             {
                 sum += array[i];
